requirements.txt
flask==2.2.3
opencv-python==4.7.0.72
numpy
pillow
gunicorn
h5py